+++ src/init.cpp
@@ -2,18 +2,32 @@
 #include <Rinternals.h>
 #include <stdlib.h> // for NULL
 #include <R_ext/Rdynload.h>
+
+#include <map>
+
 #include "systemfonts.h"
 #include "FontDescriptor.h"
 
 static ResultSet* fonts;
 
-ResultSet& get_font_list(){
+ResultSet& get_font_list() {
   return *fonts;
+}
+
+static FontReg* font_registry;
+
+FontReg& get_font_registry() {
+  return *font_registry;
 }
 
 static const R_CallMethodDef CallEntries[] = {
   {"match_font_c", (DL_FUNC) &match_font, 3},
   {"system_fonts_c", (DL_FUNC) &system_fonts, 0},
+  {"dev_string_widths_c", (DL_FUNC) &dev_string_widths, 6},
+  {"dev_string_metrics_c", (DL_FUNC) &dev_string_metrics, 6},
+  {"register_font_c", (DL_FUNC) &register_font, 3},
+  {"clear_registry_c", (DL_FUNC) &clear_registry, 0},
+  {"registry_fonts_c", (DL_FUNC) &registry_fonts, 0},
   {NULL, NULL, 0}
 };
 
@@ -22,10 +36,12 @@ extern "C" void R_init_systemfonts(DllInfo *dll) {
   R_useDynamicSymbols(dll, FALSE);
 
   fonts = new ResultSet();
+  font_registry = new FontReg();
 
   R_RegisterCCallable("systemfonts", "locate_font", (DL_FUNC)locate_font);
 }
 
 extern "C" void R_unload_systemfonts(DllInfo *dll) {
   delete fonts;
+  delete font_registry;
 }