+++ src/systemfonts.cpp
@@ -1,7 +1,11 @@
+#include <string>
+
 #include <R.h>
 #include <Rinternals.h>
+#include <R_ext/GraphicsEngine.h>
 
 #include "systemfonts.h"
+#include "utils.h"
 #include "FontDescriptor.h"
 
 // these functions are implemented by the platform
@@ -10,8 +14,50 @@ ResultSet *findFonts(FontDescriptor *);
 FontDescriptor *findFont(FontDescriptor *);
 FontDescriptor *substituteFont(char *, char *);
 
+// Default fonts based on browser behaviour
+#if defined _WIN32
+#define SANS "Arial"
+#define SERIF "Times New Roman"
+#define MONO "Courier New"
+#elif defined __APPLE__
+#define SANS "Helvetica"
+#define SERIF "Times"
+#define MONO "Courier"
+#else
+#define SANS "sans"
+#define SERIF "serif"
+#define MONO "mono"
+#endif
+
+bool locate_in_registry(const char *family, int italic, int bold, FontLoc& res) {
+  FontReg& registry = get_font_registry();
+  if (registry.empty()) return false;
+  auto search = registry.find(std::string(family));
+  if (search == registry.end()) {
+    return false;
+  }
+  int index = bold ? (italic ? 3 : 1) : (italic ? 2 : 0);
+  res.first = search->second[index].first;
+  res.second = search->second[index].second;
+  return true;
+}
+
 int locate_font(const char *family, int italic, int bold, char *path, int max_path_length) {
-  FontDescriptor font_desc(family, italic, bold);
+  FontLoc registry_match;
+  if (locate_in_registry(family, italic, bold, registry_match)) {
+    strncpy(path, registry_match.first.c_str(), max_path_length);
+    return registry_match.second;
+  }
+  
+  const char* resolved_family = family;
+  if (strcmp_no_case(family, "") || strcmp_no_case(family, "sans")) {
+    resolved_family = SANS;
+  } else if (strcmp_no_case(family, "serif")) {
+    resolved_family = SERIF;
+  } else if (strcmp_no_case(family, "mono")) {
+    resolved_family = MONO;
+  }
+  FontDescriptor font_desc(resolved_family, italic, bold);
   FontDescriptor* font_loc = findFont(&font_desc);
 
   strncpy(path, font_loc->path, max_path_length);
@@ -69,7 +115,7 @@ SEXP system_fonts() {
 
   SEXP fct_cl = PROTECT(Rf_allocVector(STRSXP, 2));
   SET_STRING_ELT(fct_cl, 0, Rf_mkChar("ordered"));
-  SET_STRING_ELT(fct_cl, 0, Rf_mkChar("factor"));
+  SET_STRING_ELT(fct_cl, 1, Rf_mkChar("factor"));
 
   SEXP weight = PROTECT(Rf_allocVector(INTSXP, n));
   SEXP weight_lvl = PROTECT(Rf_allocVector(STRSXP, 9));
@@ -139,5 +185,235 @@ SEXP system_fonts() {
   setAttrib(res, Rf_install("row.names"), row_names);
 
   UNPROTECT(16);
+  return res;
+}
+
+SEXP dev_string_widths(SEXP strings, SEXP family, SEXP face, SEXP size, SEXP cex, SEXP unit) {
+  GEUnit u;
+  switch (INTEGER(unit)[0]) {
+  case 0:
+    u = GE_CM;
+    break;
+  case 1:
+    u = GE_INCHES;
+    break;
+  case 2:
+    u = GE_DEVICE;
+    break;
+  case 3:
+    u = GE_NDC;
+    break;
+  }
+  pGEDevDesc dev = GEcurrentDevice();
+  R_GE_gcontext gc;
+  double width;
+  int n_total = LENGTH(strings);
+  int scalar_family = LENGTH(family) == 1;
+  int scalar_rest = LENGTH(face) == 1;
+  strcpy(gc.fontfamily, Rf_translateCharUTF8(STRING_ELT(family, 0)));
+  gc.fontface = INTEGER(face)[0];
+  gc.ps = REAL(size)[0];
+  gc.cex = REAL(cex)[0];
+  SEXP res = PROTECT(allocVector(REALSXP, n_total));
+  
+  for (int i = 0; i < n_total; i++) {
+    if (i > 0 && !scalar_family) {
+      strcpy(gc.fontfamily, Rf_translateCharUTF8(STRING_ELT(family, i)));
+    }
+    if (i > 0 && !scalar_rest) {
+      gc.fontface = INTEGER(face)[i];
+      gc.ps = REAL(size)[i];
+      gc.cex = REAL(cex)[i];
+    }
+    width = GEStrWidth(
+      CHAR(STRING_ELT(strings, i)), 
+      getCharCE(STRING_ELT(strings, i)), 
+      &gc, 
+      dev
+    );
+    REAL(res)[i] = GEfromDeviceWidth(width, u, dev);
+  }
+  
+  UNPROTECT(1);
+  return res;
+}
+
+SEXP dev_string_metrics(SEXP strings, SEXP family, SEXP face, SEXP size, SEXP cex, SEXP unit) {
+  GEUnit u;
+  switch (INTEGER(unit)[0]) {
+  case 0:
+    u = GE_CM;
+    break;
+  case 1:
+    u = GE_INCHES;
+    break;
+  case 2:
+    u = GE_DEVICE;
+    break;
+  case 3:
+    u = GE_NDC;
+    break;
+  }
+  pGEDevDesc dev = GEcurrentDevice();
+  R_GE_gcontext gc;
+  double width, ascent, descent;
+  int n_total = LENGTH(strings);
+  int scalar_family = LENGTH(family) == 1;
+  int scalar_rest = LENGTH(face) == 1;
+  strcpy(gc.fontfamily, Rf_translateCharUTF8(STRING_ELT(family, 0)));
+  gc.fontface = INTEGER(face)[0];
+  gc.ps = REAL(size)[0];
+  gc.cex = REAL(cex)[0];
+  SEXP w = PROTECT(allocVector(REALSXP, n_total));
+  SEXP a = PROTECT(allocVector(REALSXP, n_total));
+  SEXP d = PROTECT(allocVector(REALSXP, n_total));
+  
+  for (int i = 0; i < n_total; i++) {
+    if (i > 0 && !scalar_family) {
+      strcpy(gc.fontfamily, Rf_translateCharUTF8(STRING_ELT(family, i)));
+    }
+    if (i > 0 && !scalar_rest) {
+      gc.fontface = INTEGER(face)[i];
+      gc.ps = REAL(size)[i];
+      gc.cex = REAL(cex)[i];
+    }
+    GEStrMetric(
+      CHAR(STRING_ELT(strings, i)), 
+      getCharCE(STRING_ELT(strings, i)), 
+      &gc,
+      &ascent, &descent, &width,
+      dev
+    );
+    REAL(w)[i] = GEfromDeviceWidth(width, u, dev);
+    REAL(a)[i] = GEfromDeviceWidth(ascent, u, dev);
+    REAL(d)[i] = GEfromDeviceWidth(descent, u, dev);
+  }
+  
+  SEXP res = PROTECT(allocVector(VECSXP, 3));
+  SET_VECTOR_ELT(res, 0, w);
+  SET_VECTOR_ELT(res, 1, a);
+  SET_VECTOR_ELT(res, 2, d);
+  
+  SEXP row_names = PROTECT(Rf_allocVector(REALSXP, 2));
+  REAL(row_names)[0] = NA_REAL;
+  REAL(row_names)[1] = -n_total;
+  setAttrib(res, Rf_install("row.names"), row_names);
+  
+  SEXP names = PROTECT(Rf_allocVector(STRSXP, 3));
+  SET_STRING_ELT(names, 0, Rf_mkChar("width"));
+  SET_STRING_ELT(names, 1, Rf_mkChar("ascent"));
+  SET_STRING_ELT(names, 2, Rf_mkChar("descent"));
+  setAttrib(res, Rf_install("names"), names);
+  
+  SEXP cl = PROTECT(Rf_allocVector(STRSXP, 3));
+  SET_STRING_ELT(cl, 0, Rf_mkChar("tbl_df"));
+  SET_STRING_ELT(cl, 1, Rf_mkChar("tbl"));
+  SET_STRING_ELT(cl, 2, Rf_mkChar("data.frame"));
+  Rf_classgets(res, cl);
+  
+  UNPROTECT(7);
+  return res;
+}
+
+SEXP register_font(SEXP family, SEXP paths, SEXP indices) {
+  FontReg& registry = get_font_registry();
+  std::string name = Rf_translateCharUTF8(STRING_ELT(family, 0));
+  FontCollection col;
+  for (int i = 0; i < LENGTH(paths); i++) {
+    std::string font_path = Rf_translateCharUTF8(STRING_ELT(paths, i));
+    FontLoc font(font_path, INTEGER(indices)[i]);
+    col.push_back(font);
+  }
+  registry[name] = col;
+  
+  return R_NilValue;
+}
+
+SEXP clear_registry() {
+  FontReg& registry = get_font_registry();
+  registry.clear();
+  
+  return R_NilValue;
+}
+
+SEXP registry_fonts() {
+  FontReg& registry = get_font_registry();
+  int n_reg = registry.size();
+  int n = n_reg * 4;
+  
+  SEXP res = PROTECT(Rf_allocVector(VECSXP, 6));
+  
+  SEXP cl = PROTECT(Rf_allocVector(STRSXP, 3));
+  SET_STRING_ELT(cl, 0, Rf_mkChar("tbl_df"));
+  SET_STRING_ELT(cl, 1, Rf_mkChar("tbl"));
+  SET_STRING_ELT(cl, 2, Rf_mkChar("data.frame"));
+  Rf_classgets(res, cl);
+  
+  SEXP names = PROTECT(Rf_allocVector(STRSXP, 6));
+  SET_STRING_ELT(names, 0, Rf_mkChar("path"));
+  SET_STRING_ELT(names, 1, Rf_mkChar("index"));
+  SET_STRING_ELT(names, 2, Rf_mkChar("family"));
+  SET_STRING_ELT(names, 3, Rf_mkChar("style"));
+  SET_STRING_ELT(names, 4, Rf_mkChar("weight"));
+  SET_STRING_ELT(names, 5, Rf_mkChar("italic"));
+  setAttrib(res, Rf_install("names"), names);
+  
+  SEXP path = PROTECT(Rf_allocVector(STRSXP, n));
+  SEXP index = PROTECT(Rf_allocVector(INTSXP, n));
+  SEXP family = PROTECT(Rf_allocVector(STRSXP, n));
+  SEXP style = PROTECT(Rf_allocVector(STRSXP, n));
+  
+  SEXP fct_cl = PROTECT(Rf_allocVector(STRSXP, 2));
+  SET_STRING_ELT(fct_cl, 0, Rf_mkChar("ordered"));
+  SET_STRING_ELT(fct_cl, 1, Rf_mkChar("factor"));
+  
+  SEXP weight = PROTECT(Rf_allocVector(INTSXP, n));
+  SEXP weight_lvl = PROTECT(Rf_allocVector(STRSXP, 2));
+  SET_STRING_ELT(weight_lvl, 0, Rf_mkChar("normal"));
+  SET_STRING_ELT(weight_lvl, 1, Rf_mkChar("bold"));
+  Rf_classgets(weight, fct_cl);
+  Rf_setAttrib(weight, Rf_install("levels"), weight_lvl);
+  
+  SEXP italic = PROTECT(Rf_allocVector(LGLSXP, n));
+  
+  SET_VECTOR_ELT(res, 0, path);
+  SET_VECTOR_ELT(res, 1, index);
+  SET_VECTOR_ELT(res, 2, family);
+  SET_VECTOR_ELT(res, 3, style);
+  SET_VECTOR_ELT(res, 4, weight);
+  SET_VECTOR_ELT(res, 5, italic);
+  
+  int i = 0;
+  for (auto it = registry.begin(); it != registry.end(); ++it) {
+    for (int j = 0; j < 4; j++) {
+      SET_STRING_ELT(path, i, Rf_mkChar(it->second[j].first.c_str()));
+      INTEGER(index)[i] = it->second[j].second;
+      SET_STRING_ELT(family, i, Rf_mkChar(it->first.c_str()));
+      switch (j) {
+      case 0: 
+        SET_STRING_ELT(style, i, Rf_mkChar("Regular"));
+        break;
+      case 1:
+        SET_STRING_ELT(style, i, Rf_mkChar("Bold"));
+        break;
+      case 2:
+        SET_STRING_ELT(style, i, Rf_mkChar("Italic"));
+        break;
+      case 3:
+        SET_STRING_ELT(style, i, Rf_mkChar("Bold Italic"));
+        break;
+      }
+      INTEGER(weight)[i] = 1 + (int) (j == 1 || j == 3);
+      INTEGER(italic)[i] = (int) (j > 1);
+      i++;
+    }
+  }
+  
+  SEXP row_names = PROTECT(Rf_allocVector(REALSXP, 2));
+  REAL(row_names)[0] = NA_REAL;
+  REAL(row_names)[1] = -n;
+  setAttrib(res, Rf_install("row.names"), row_names);
+  
+  UNPROTECT(11);
   return res;
 }